resume_parser_starter/requirements.txt
Flask
Flask-SQLAlchemy
sqlalchemy
psycopg2-binary
pdfplumber
python-docx
phonenumbers
rapidfuzz
